Add files via upload
This is the modularized and parameterized CNN script with a simple call to go along with it. Feel free to mess with the parameters to see if you can get a better result. 

This script also conditionally supports the use of either CPU or GPU. But I haven't tested it with a GPU yet.